Reduz RAM de 646B para 587B (branch tiny)
Meta audaciosa: aproximar do firmware original da MegaSquirt MS1
(512B, em assembly). Cortes aplicados sem afetar timing crítico:

- Remove struct FuelSchedule morta (scheduler.h/.cpp): 3 instâncias +
  setFuelSchedule()/clearFuelSchedule() nunca chamadas - injeção real
  já usa InjectorPollingState via polling.
- Remove struct Table2D morta (tables.h/.cpp): zero instâncias.
- Remove cache de interpolação do Table3D (lastX/lastY/lastInputX/
  lastInputY/lastOutput, 16B): getTableValue() recalcula a cada
  lookup em vez de cachear; busca de índice sempre parte do zero.
  Custo é CPU extra, não corretude.
- Downcast TriggerState.toothTotalCount/triggerActualTeeth/
  toothCurrentCount de uint16_t para uint8_t (faixa real max ~120
  dentes).
- Reduz SERIAL_RX/TX_BUFFER_SIZE de 16 para 8 no platformio.ini.
- Adiciona .pio/.pio_venv/simulator/.pio ao .gitignore.

Build validado limpo nos envs uno/nanoatmega328/megaatmega2560.

+++ slowduino/decoders.h
@@ -25,9 +25,9 @@ struct TriggerState {
   volatile uint32_t toothOneTime;          // Timestamp do dente #1 (referência)
 
   // Contadores
-  volatile uint16_t toothCurrentCount;     // Contador de dentes na revolução atual
-  volatile uint16_t toothTotalCount;       // Total de dentes na roda (ex: 36)
-  volatile uint16_t triggerActualTeeth;    // Dentes reais (ex: 35 em 36-1)
+  volatile uint8_t toothCurrentCount;      // Contador de dentes na revolução atual
+  volatile uint8_t toothTotalCount;        // Total de dentes na roda (ex: 36)
+  volatile uint8_t triggerActualTeeth;     // Dentes reais (ex: 35 em 36-1)
 
   // Gaps e sincronização
   volatile uint32_t curGap;                // Gap atual entre dentes
